requirements.txt
python-dotenv==0.21.0
Flask==2.2.2
Mastodon.py==1.6.2
zappa==0.55.0